+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [4.2.0] - 2022-01-20
+### Added
+- Added Secure User ID Store adapter
+- Send "new user" email notification to user's HR contact upon creation in the IdP
+### Changed
+- Renamed internal user fields from "broker" to "internal"
+- Added two new user fields: `hr_contact_name`, and `hr_contact_email`
 
 ## [4.1.4] - 2021-11-10
 ### Removed